Make SchemaService the one schema interface, as a class
The engine went through SchemaEnforcement while the AI layer called
SchemaService directly, so the same feature had two entry points and only
one of them was metered. Both now call SchemaService.validate, which is
also the shape the commercial server uses.

SchemaService is a class rather than an interface with a single
implementation. There is only ever one registry, and the seam worth
substituting per backend is SchemaDAO below it. The public method set is
unchanged at eleven.

SchemaEnforcement is deleted and its gate inlined at the five enforcement
points. The payload is still supplied lazily: WorkflowModel.getInput and
getOutput merge the inline and external-storage maps and reset the payload
field, so reading one for an execution that never opted in would do that
work on every execution.

Metering moves into validate, the single point every caller passes
through. Two consequences worth noting for anyone with a dashboard:
schema_validation_failure loses its boundary and workflowName tags and
schema_validation loses boundary, because validate is not told which
boundary called it; and LLM output validation is now counted too, where
before only the engine was. The recorded reason also loses its "Task step
output: " prefix, matching the commercial server.
